requirements.txt
pyserial
Flask
Flask-CORS
boto3
python-dotenv